# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Travel Agent API
  description: API for conversational flight search with price monitoring and alerts
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.travelagent.example.com/v1
    description: Production server
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logged out successfully
...
  /preferences:
    get:
      tags:
        - Preferences
      summary: Get user preferences
      operationId: getPreferences
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
    
    patch:
      tags:
        - Preferences
      summary: Update preferences
      operationId: updatePreferences
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesUpdateRequest'
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
  /preferences/reset:
    post:
      tags:
        - Preferences
      summary: Reset preferences to defaults
      operationId: resetPreferences
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Preferences reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            customSessionId:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    PreferencesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      properties:
        userId:
          type: string
        notifications:
          type: object
          properties:
            emailEnabled:
              type: boolean
            frequency:
              type: string
              enum: [INSTANT, HOURLY, DAILY, WEEKLY, NEVER]
            priceAlerts:
              type: object
            searchExpiration:
              type: object
            marketing:
              type: object
        search:
          type: object
          properties:
            defaultCabinClass:
              type: string
            preferredAirlines:
              type: array
              items:
                type: string
            maxStops:
              type: integer
        display:
          type: object
          properties:
            currency:
              type: string
            dateFormat:
              type: string
            timeFormat:
              type: string
            theme:
              type: string
    PreferencesUpdateRequest:
      type: object
      properties:
        notifications:
          type: object
        search:
          type: object
        display:
          type: object
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'